+++ Policy-Service/src/main/java/com/project/service/PolicyService.java
@@ -5,11 +5,11 @@ import java.util.List;
 import com.project.model.Policy;
 
 public interface PolicyService {
-    Policy createPolicy(Policy policy, Long customerId, Long agentId);
-    Policy updatePolicy(Long policyId, Policy updatedPolicy);
-    void deletePolicy(Long policyId);
-    Policy getPolicyById(Long policyId);
+    Policy createPolicy(Policy policy, Integer customerId, Integer agentId);
+    Policy updatePolicy(Integer policyId, Policy updatedPolicy);
+    void deletePolicy(Integer policyId);
+    Policy getPolicyById(Integer policyId);
     List<Policy> getAllPolicies();
-    List<Policy> getPoliciesByCustomer(Long customerId);
-    List<Policy> getPoliciesByAgent(Long agentId);
+    List<Policy> getPoliciesByCustomer(Integer customerId);
+    List<Policy> getPoliciesByAgent(Integer agentId);
 }
